Change structure of components
