+++ src/components/family-tree/canvas/canvas.scss
@@ -9,12 +9,4 @@
     cursor: grabbing;
     user-select: none;
   }
-
-  &__members-container {
-    margin: auto;
-    display: flex;
-    justify-content: center;
-    padding: 40px;
-    box-sizing: border-box;
-  }
 }
