+++ src/components/family-tree/familyTree.scss
@@ -1,0 +1,7 @@
+.members-container {
+  margin: auto;
+  display: flex;
+  justify-content: center;
+  padding: 40px;
+  box-sizing: border-box;
+}
