+++ src/components/family-tree/member/member-info/details/details.scss
