+++ src/components/family-tree/member/member-info/general-info/generalInfo.js
