+++ src/components/family-tree/member/member-info/general-info/picture/picture.scss
