+++ src/components/family-tree/member/member-info/memberInfo.js
