+++ src/components/family-tree/member/member.scss
