+++ src/components/family-tree/placeholder/placeholder.scss
