+++ src/components/family-tree/search/result/result.js
