+++ src/components/family-tree/search/result/result.scss
