+++ src/components/family-tree/search/search.scss
